+++ src/sdk/java/sql-jdbc/src/main/java/com/_4paradigm/hybridsql/fedb/jdbc/SQLInsertMetaData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com._4paradigm.hybridsql.jdbc;
+package com._4paradigm.hybridsql.fedb.jdbc;
 
 import com._4paradigm.hybridsql.DataType;
 import com._4paradigm.hybridsql.Schema;
@@ -22,13 +22,19 @@ import com._4paradigm.hybridsql.Schema;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.Types;
+import java.util.List;
 
-public class SQLResultSetMetaData implements ResultSetMetaData {
+public class SQLInsertMetaData implements ResultSetMetaData {
 
-    private Schema schema;
-
-    public SQLResultSetMetaData(Schema schema) {
+    private List<DataType> schema;
+    private Schema realSchema;
+    private List<Integer> idx;
+    public SQLInsertMetaData(List<DataType> schema,
+                             Schema realSchema,
+                             List<Integer> idx) {
         this.schema = schema;
+        this.realSchema = realSchema;
+        this.idx = idx;
     }
 
     private void checkSchemaNull() throws SQLException {
@@ -41,7 +47,7 @@ public class SQLResultSetMetaData implements ResultSetMetaData {
         if (i <= 0) {
             throw new SQLException("index underflow");
         }
-        if (i > schema.GetColumnCnt()) {
+        if (i > schema.size()) {
             throw new SQLException("index overflow");
         }
     }
@@ -78,7 +84,7 @@ public class SQLResultSetMetaData implements ResultSetMetaData {
     @Override
     public int getColumnCount() throws SQLException {
         checkSchemaNull();
-        return schema.GetColumnCnt();
+        return schema.size();
     }
 
     @Override
@@ -108,7 +114,8 @@ public class SQLResultSetMetaData implements ResultSetMetaData {
     @Override
     public int isNullable(int i) throws SQLException {
         check(i);
-        if (schema.IsColumnNotNull(i - 1)) {
+        int index = idx.get(i - 1);
+        if (realSchema.IsColumnNotNull(index)) {
             return columnNoNulls;
         } else {
             return columnNullable;
@@ -136,7 +143,8 @@ public class SQLResultSetMetaData implements ResultSetMetaData {
     @Override
     public String getColumnName(int i) throws SQLException {
         check(i);
-        return schema.GetColumnName(i - 1);
+        int index = idx.get(i - 1);
+        return realSchema.GetColumnName(index);
     }
 
     @Override
@@ -172,7 +180,7 @@ public class SQLResultSetMetaData implements ResultSetMetaData {
     @Override
     public int getColumnType(int i) throws SQLException {
         check(i);
-        DataType dataType = schema.GetColumnType(i - 1);
+        DataType dataType = schema.get(i - 1);
         return type2SqlType(dataType);
     }
