+++ src/sdk/java/sql-jdbc/src/main/java/com/_4paradigm/hybridsql/fedb/sdk/Common.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com._4paradigm.hybridsql.sdk;
+package com._4paradigm.hybridsql.fedb.sdk;
 
 import com._4paradigm.hybridsql.DataType;
 import com._4paradigm.hybridsql.Schema;
@@ -49,7 +49,7 @@ public class Common {
         }
     }
 
-    public static com._4paradigm.hybridsql.sdk.Schema convertSchema(Schema schema) throws SQLException {
+    public static Schema convertSchema(Schema schema) throws SQLException {
         if (schema == null || schema.GetColumnCnt() == 0) {
             throw new SQLException("schema is null or empty");
         }
@@ -62,6 +62,6 @@ public class Common {
             column.setConstant(schema.IsConstant(i));
             columnList.add(column);
         }
-        return new com._4paradigm.hybridsql.sdk.Schema(columnList);
+        return new Schema(columnList);
     }
 }
