+++ src/sdk/java/sql-jdbc/src/main/java/com/_4paradigm/hybridsql/fedb/sdk/impl/CallablePreparedStatementImpl.java
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package com._4paradigm.hybridsql.sdk.impl;
+package com._4paradigm.hybridsql.fedb.sdk.impl;
 
 import com._4paradigm.hybridsql.QueryFuture;
 import com._4paradigm.hybridsql.SQLRouter;
 import com._4paradigm.hybridsql.Status;
-import com._4paradigm.hybridsql.jdbc.CallablePreparedStatement;
-import com._4paradigm.hybridsql.jdbc.SQLResultSet;
+import com._4paradigm.hybridsql.fedb.jdbc.CallablePreparedStatement;
+import com._4paradigm.hybridsql.fedb.jdbc.SQLResultSet;
+import com._4paradigm.hybridsql.fedb.sdk.QueryFuture;
 
 import java.sql.SQLException;
 import java.util.concurrent.TimeUnit;
@@ -54,7 +55,7 @@ public class CallablePreparedStatementImpl extends CallablePreparedStatement {
     }
 
     @Override
-    public com._4paradigm.hybridsql.sdk.QueryFuture executeQueryAsync(long timeOut, TimeUnit unit) throws SQLException {
+    public QueryFuture executeQueryAsync(long timeOut, TimeUnit unit) throws SQLException {
         checkClosed();
         dataBuild();
         Status status = new Status();
@@ -68,7 +69,7 @@ public class CallablePreparedStatementImpl extends CallablePreparedStatement {
             throw new SQLException("call procedure fail, msg: " + msg);
         }
         status.delete();
-        return new com._4paradigm.hybridsql.sdk.QueryFuture(queryFuture);
+        return new QueryFuture(queryFuture);
     }
 
 }
